Student_Management-master/requirements_selenium.txt
# Dependencias para pruebas con Selenium
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0

# Si necesitas estas dependencias adicionales
pytest>=7.4.0
pytest-django>=4.5.0
pytest-xvfb>=3.0.0  # Para ejecutar en servidores sin GUI